internal: use unsigned types where relevant and enable -Wsign-conversion

+++ libnopegl/src/api.c
@@ -287,9 +287,9 @@ int ngli_ctx_set_scene(struct ngl_ctx *s, struct ngl_scene *scene)
     }
 
     // Re-compute the viewport according to the new scene aspect ratio
-    int32_t width, height;
+    uint32_t width, height;
     ngpu_ctx_get_default_rendertarget_size(s->gpu_ctx, &width, &height);
-    s->viewport = compute_scene_viewport(s->scene, width, height);
+    s->viewport = compute_scene_viewport(s->scene, (int32_t)width, (int32_t)height);
     s->scissor = (struct ngpu_scissor){0, 0, width, height};
 
     const struct ngl_config *config = &s->config;
@@ -420,13 +420,13 @@ fail:
     return ret;
 }
 
-int ngli_ctx_resize(struct ngl_ctx *s, int32_t width, int32_t height)
+int ngli_ctx_resize(struct ngl_ctx *s, uint32_t width, uint32_t height)
 {
     int ret = ngpu_ctx_resize(s->gpu_ctx, width, height);
     if (ret < 0)
         return ret;
 
-    s->viewport = compute_scene_viewport(s->scene, width, height);
+    s->viewport = compute_scene_viewport(s->scene, (int32_t)width, (int32_t)height);
     s->scissor = (struct ngpu_scissor){0, 0, width, height};
 
     return 0;
@@ -773,7 +773,7 @@ void ngl_reset_backend(struct ngl_backend *backend)
     backend_reset(backend);
 }
 
-int ngl_resize(struct ngl_ctx *s, int32_t width, int32_t height)
+int ngl_resize(struct ngl_ctx *s, uint32_t width, uint32_t height)
 {
     if (!s->configured) {
         LOG(ERROR, "context must be configured before resizing rendering buffers");
